fix(test): add web tests
