--- tests/web/login_test.go
@@ -1,23 +1,0 @@
-package web
-
-import (
-	"github.com/stretchr/testify/assert"
-	"strings"
-	"testing"
-	"time"
-)
-
-func TestLogin(t *testing.T) {
-	defer StopDriverOnPanic(t)
-
-	assert.Equal(t, page.Navigate(url("/admin")), nil)
-	assert.Equal(t, page.Find("#username").Fill("admin"), nil)
-	assert.Equal(t, page.Find("#password").Fill("admin"), nil)
-	assert.Equal(t, page.FindByButton("login").Click(), nil)
-
-	time.Sleep(time.Second * 3)
-
-	content, err := page.HTML()
-	assert.Equal(t, err, nil)
-	assert.Equal(t, strings.Contains(content, "main-header"), true)
-}
